split up expr for rewriting, visiting, and simplification
